Refactor microkernel headers
- Expose a unified `api.h` header.
- Fold separate `_types.h` headers.
- Fold some internal build targets.
- Caught some stray `#include <assert.h>`.
