+++ runtime/src/iree/builtins/ukernel/arch/arm_64/CMakeLists.txt
@@ -85,7 +85,7 @@ iree_cc_library(
   DEPS
     iree::base::core_headers
     iree::schemas::cpu_data
-    iree::builtins::ukernel::common
+    iree::builtins::ukernel::headers
     ${IREE_UK_MMT4D_TILE_ARM_64_DEPS}
   PUBLIC
 )
@@ -119,7 +119,7 @@ iree_cc_library(
   DEPS
     iree::base::core_headers
     iree::schemas::cpu_data
-    iree::builtins::ukernel::common
+    iree::builtins::ukernel::headers
     ::pack_tile_arm_64
   PUBLIC
 )
@@ -138,6 +138,6 @@ iree_cc_library(
   DEPS
     iree::base::core_headers
     iree::schemas::cpu_data
-    iree::builtins::ukernel::common
+    iree::builtins::ukernel::headers
   PUBLIC
 )
