+++ runtime/src/iree/builtins/ukernel/arch/arm_64/pack_arm_64.h
@@ -7,7 +7,7 @@
 #ifndef IREE_BUILTINS_UKERNEL_ARCH_ARM_64_PACK_ARM_64_H_
 #define IREE_BUILTINS_UKERNEL_ARCH_ARM_64_PACK_ARM_64_H_
 
-#include "iree/builtins/ukernel/pack_types.h"
+#include "iree/builtins/ukernel/pack.h"
 
 // Returns the arm64 tile function to use for the pack op with given params, or
 // NULL if no suitable arm64 tile function exists for these params, in which
