+++ runtime/src/iree/builtins/ukernel/arch/arm_64/query_tile_sizes_arm_64.h
@@ -7,7 +7,7 @@
 #ifndef IREE_BUILTINS_UKERNEL_ARCH_ARM_64_QUERY_TILE_SIZES_ARM_64_H_
 #define IREE_BUILTINS_UKERNEL_ARCH_ARM_64_QUERY_TILE_SIZES_ARM_64_H_
 
-#include "iree/builtins/ukernel/query_tile_sizes_types.h"
+#include "iree/builtins/ukernel/query_tile_sizes.h"
 
 bool iree_uk_query_matmul_tile_sizes_arm_64(
     const iree_uk_query_tile_sizes_2d_params_t* params,
