+++ runtime/src/iree/modules/vmvx/module.c
@@ -19,11 +19,7 @@
 // Include the ukernel support library so that we can use its implementations
 // as fixed-function components of the runtime.
 #include "iree/base/internal/cpu.h"
-#include "iree/builtins/ukernel/elementwise.h"
-#include "iree/builtins/ukernel/mmt4d.h"
-#include "iree/builtins/ukernel/pack.h"
-#include "iree/builtins/ukernel/query_tile_sizes.h"
-#include "iree/builtins/ukernel/unpack.h"
+#include "iree/builtins/ukernel/api.h"
 
 #define IREE_VMVX_MODULE_VERSION_0_0 0x00000000u
 #define IREE_VMVX_MODULE_VERSION_LATEST IREE_VMVX_MODULE_VERSION_0_0
